Move add_provenance_for_stmts to parent class.
